sesion_06/requirements.txt
Django==3.1.3
graphene-django==2.13.0
pyjwt==1.7.1
